refactored code and moved the audio into a more global scope, now maches class diagramm
