+++ auditory_stimulation/auditory_tagging/noise_tagging_tagger.py
@@ -3,7 +3,8 @@ from typing import List, Tuple, Optional
 import numpy as np
 import numpy.typing as npt
 
-from auditory_stimulation.auditory_tagging.auditory_tagger import AAudioTagger, Audio, AAudioTaggerFactory
+from auditory_stimulation.audio import Audio
+from auditory_stimulation.auditory_tagging.auditory_tagger import AAudioTagger, AAudioTaggerFactory
 
 Code = npt.NDArray[np.int16]
