+++ tests/auditory_tagging/helper/sound_players_test.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pytest
 
-from auditory_stimulation.auditory_tagging.auditory_tagger import Audio
+from auditory_stimulation.audio import Audio
 from auditory_stimulation.view.sound_players import psychopy_player
